update header details
