+++ src/messages/GE.json
@@ -439,5 +439,6 @@
   "SDKs documentation": "To read more about our SDKs, go to the documentation page at ",
   "Settings": "Settings",
   "sign in other account": "Sign in with a different account",
-  "Sample Tenant": "Sample Tenant"
+  "Sample Tenant": "Sample Tenant",
+  "Help": "Help"
 }
